+++ CHANGES.md
@@ -4,6 +4,16 @@ CHANGELOG
 8.1.0+dev  (XXXX-XX-XX)                
 -----------------------
 
+**Bug fixes**
+
+- Fix Log Entry GeoJSON with no geometry
+- Fix log entry access by creating mapentity.read_logentry permission
+- Hide log entry menu for users without permission.
+- Hide "Show full history" from object for users without permission.
+- Hide admin menu entry for users without permission.
+- Fix message in default in 404 view
+- Fix exception recursion in default error 500 view
+
 
 8.1.0      (2022-06-03)                
 -----------------------